requirements.txt
fastapi
uvicorn[standard]
requests
openai
python-dotenv
airportsdata
httpx
asyncpg>=0.29.0
redis>=5.0.0
databases[postgresql]>=0.9.0